requirements.txt
openai==1.54.3
python-dotenv
pydantic-ai-slim[openai]
pydantic-ai[logfire]
pydantic
gunicorn
logfire
flask
